GUISE-155, GUISE-161: Removed footer padding.
The benefit of this is uncertain, or at least the pressing need is less
clear than it was before. It now makes less sense that all footers would
have padding; a hero or card may not require padding, and if padding is
needed the footer element can indicate the new panel component.